- fixed out-of-bounds memory accesses in nvBWT - added possibility to specify genome index to sufsort-test
[git-p4: depot-paths = "//sw/gpgpu/nvbio/main/": change = 18185737]

+++ nvbio/sufsort/sufsort_utils.h
@@ -350,7 +350,7 @@ struct StringBWTHandler
     uint32                          primary;
     uint32                          n_output;
     output_iterator                 output;
-    thrust::device_vector<uint32>   d_block_bwt;
+    thrust::device_vector<uint8>    d_block_bwt;
 };
 
 /// a utility StringSuffixHandler to retain a Sampled Suffix Array
